+++ packages/sample-app/sample_app/langchain_agent.py
@@ -1,8 +1,8 @@
-from langchain.agents import AgentType
-from langchain.agents import initialize_agent, Tool
-from langchain.chains import LLMMathChain
+from langchain_classic.agents import AgentType
+from langchain_classic.agents import initialize_agent, Tool
+from langchain_classic.chains import LLMMathChain
 from langchain_community.llms.openai import OpenAI
-from langchain.utilities import DuckDuckGoSearchAPIWrapper
+from langchain_community.utilities import DuckDuckGoSearchAPIWrapper
 
 
 from traceloop.sdk import Traceloop